Creating Testing Branch
Hey, didn't want to mess with the code you had, but I also didn't want to fork it into a separate repo either. I just created a branch so I can mess with things but not disrupt what is already working for Q1-5

I also named the file I was working with main2 just as to not have 2 files with the same name, we can clean up the files and delete redundancies later.