Fixed Q6
because astar is keeping track of the trajectory the total we were getting was already cumulative, but we were also keeping a total variable, so we were getting a cumulative sum of the cumulative sum and it was growing out of control really fast...

also added a couple lines of code to reset it back to 0 for this question.

Going to finish q7 tonight, saw your notes that you modified the code for that question, thanks a lot!